+++ src/advanced-actions.ts
@@ -1,23 +1,23 @@
 import { Project } from './project';
-import * as vscode from 'vscode';
 import { PackageManager, npmInstall } from './node-commands';
 import { confirm, getRunOutput, isWindows, replaceAll } from './utilities';
 import { write, writeError, writeIonic } from './logging';
 import { isGreaterOrEqual, isLess } from './analyzer';
-import { readAngularJson, writeAngularJson } from './rules-angular-json';
-import { join } from 'path';
-import { existsSync, readFileSync, writeFileSync } from 'fs';
+import { fixGlobalScss, readAngularJson, writeAngularJson } from './rules-angular-json';
 import { ionicState } from './ionic-tree-provider';
 import { clearIgnored } from './ignore';
-import { CommandName } from './command-name';
+import { CommandName, InternalCommand } from './command-name';
+import { ProgressLocation, commands, window } from 'vscode';
+import { sep } from 'path';
 
 enum Features {
   migrateToPNPM = '$(find-replace) Migrate to PNPM',
   migrateToNX = '$(outline-view-icon) Migrate to NX',
   reinstallNodeModules = '$(extensions-sync-enabled) Reinstall Node Modules',
   angularESBuild = '$(test-view-icon) Switch from WebPack to ESBuild (experimental)',
+  migrateAngularControlFlow = '$(test-view-icon) Migrate to the built-in control flow syntax',
   showIgnoredRecommendations = '$(light-bulb) Show Ignored Recommendations',
-  migrateAngularStandalone = '$(find-replace) Migrate to Ionic standalone components',
+  migrateAngularStandalone = '$(test-view-icon) Migrate to Ionic standalone components',
 }
 
 export async function advancedActions(project: Project) {
@@ -31,24 +31,32 @@ export async function advancedActions(project: Project) {
     if (isGreaterOrEqual('@angular/core', '14.0.0')) {
       picks.push(Features.migrateAngularStandalone);
     }
+    if (isGreaterOrEqual('@angular/core', '17.0.0')) {
+      picks.push(Features.migrateAngularControlFlow);
+    }
     picks.push(Features.reinstallNodeModules);
     picks.push(Features.showIgnoredRecommendations);
   }
   if (isGreaterOrEqual('@angular-devkit/build-angular', '14.0.0')) {
-    if (!angularUsingESBuild(project)) {
-      picks.push(Features.angularESBuild);
+    if (!isGreaterOrEqual('@angular/core', '17.0.0')) {
+      if (!angularUsingESBuild(project)) {
+        picks.push(Features.angularESBuild);
+      }
     }
   }
-  const selection = await vscode.window.showQuickPick(picks, {});
+  const selection = await window.showQuickPick(picks, {});
   switch (selection) {
     case Features.migrateToPNPM:
       await runCommands(migrateToPNPM(), selection, project);
       break;
     case Features.migrateToNX:
-      await vscode.window.showInformationMessage('Run the following command: npx nx init', 'OK');
+      await window.showInformationMessage('Run the following command: npx nx init', 'OK');
       break;
     case Features.reinstallNodeModules:
       await runCommands(reinstallNodeModules(), selection, project);
+      break;
+    case Features.migrateAngularControlFlow:
+      migrateAngularControlFlow(selection, project);
       break;
     case Features.angularESBuild:
       switchAngularToESBuild(project);
@@ -64,6 +72,19 @@ export async function advancedActions(project: Project) {
 
 function migrateToPNPM(): Array<string> {
   return ['pnpm -v', 'rm -rf node_modules', 'pnpm import', 'pnpm install', 'rm package-lock.json'];
+}
+
+async function migrateAngularControlFlow(selection: string, project: Project) {
+  if (
+    !(await confirm(
+      'This will change your Angular templates to use the new built-in control flow syntax. Are you sure?',
+      'Continue'
+    ))
+  )
+    return;
+
+  const commands = [`npx ng generate @angular/core:control-flow --interactive=false --defaults=true --path=".${sep}"`];
+  await runCommands(commands, selection, project);
 }
 
 async function migrateToAngularStandalone(selection: string, project: Project) {
@@ -99,20 +120,17 @@ function reinstallNodeModules(): Array<string> {
 
 function showIgnoredRecommendations(): void {
   clearIgnored(ionicState.context);
-  vscode.commands.executeCommand(CommandName.Refresh);
+  commands.executeCommand(CommandName.Refresh);
 }
 
-async function runCommands(commands: Array<string>, title: string, project: Project): Promise<void> {
+export async function runCommands(commands: Array<string>, title: string, project: Project): Promise<void> {
   try {
     if (title.includes(')')) {
       title = title.substring(title.indexOf(')') + 1);
     }
-    await vscode.window.withProgress(
-      { location: vscode.ProgressLocation.Notification, title, cancellable: false },
-      async () => {
-        await run(commands, project.folder);
-      }
-    );
+    await window.withProgress({ location: ProgressLocation.Notification, title, cancellable: false }, async () => {
+      await run(commands, project.folder);
+    });
 
     writeIonic(`Completed ${title}`);
   } catch (err) {
@@ -122,11 +140,11 @@ async function runCommands(commands: Array<string>, title: string, project: Proj
 
 async function run(commands: Array<string>, folder: string): Promise<void> {
   for (const command of commands) {
-    writeIonic(command);
+    writeIonic(replaceAll(command, InternalCommand.cwd, ''));
     try {
       write(await getRunOutput(command, folder));
     } catch (err) {
-      writeError(err);
+      //writeError(err);
       break;
     }
   }
@@ -159,20 +177,6 @@ function switchAngularToESBuild(project: Project): void {
   if (changes) {
     fixGlobalScss(project);
     writeAngularJson(project, angular);
-    vscode.window.showInformationMessage(`The Angular project has been changed to esbuild. Enjoy faster builds!`, 'OK');
-  }
-}
-
-function fixGlobalScss(project: Project) {
-  try {
-    const filename = join(project.folder, 'src', 'global.scss');
-    if (existsSync(filename)) {
-      let txt = readFileSync(filename, 'utf8');
-      txt = replaceAll(txt, `@import "~@ionic/`, `@import "@ionic/`);
-      txt = replaceAll(txt, `@import '~@ionic/`, `@import '@ionic/`);
-      writeFileSync(filename, txt);
-    }
-  } catch (error) {
-    writeError(`Unable to write global.scss ${error}`);
+    window.showInformationMessage(`The Angular project has been changed to esbuild. Enjoy faster builds!`, 'OK');
   }
 }