+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### Version 1.59
+
+- Add control flow migration for Angular 17 projects
+- Add recommendation to switch to ESBuild for Angular 17 projects
+- Add recommendation for migrations between major Angular versions
+- Add rrecommendation to switch Angular CLI Package Manager if using pnpm or yarn
+- Ionic CLI updated from 7.1.1 to 7.1.5
+- Fix .browserslistrc getting mangled
+- Fix error output being repeated twice for some commands
+- Fix pnpm install node modules
+
 ### Version 1.58
 
 - Fix to update ionicons to > 7.2.1 when migrating to standalone components